refactor : functions made to utils
